pyproject.toml
[project]
name = "englajimmy-backend"
version = "0.1.0"
description = "Backend for Englajimmy"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic[email]>=2.0",
    "psycopg2-binary>=2.9",
]

[tool.uv]
